Support multiple stanby node mode on standby side.
We introduce new functions such as indirect_polling(), indirect_kill()
which are used for idirectly polling and updating standby's local cache.

In most case, we don't want to promote any standby server as long as
the master server seems to be live. If the standby server checks connection
to master only from itself, it would not be work fine in case of network failuer
between master server and that standby server. So we introduce indirectly polling
mechanism which asks other standby server to poll the master server. If one
standby can not connect to master but other standby says we can connect to master
server, it's OK. On the other hand if the all stnadbys say that we could
not connect to the master server, it's time to promote.

The master server can update its local cache information when adding new
node, but standby server could not know that management table is updated.
So we introduce indirectly sending signal function which sends SIGUSR1 signal
to pg_keeper process via SQL. Using this function, we can probagate the signal
to all other standbys.

Also we change the process title so that it has the number of standbys on its
local cache.

Masahiko Sawada

+++ standby.c
@@ -11,6 +11,7 @@
 
 #include "pg_keeper.h"
 #include "syncrep.h"
+#include "util.h"
 
 /* These are always necessary for a bgworker */
 #include "access/xlog.h"
@@ -30,14 +31,15 @@
 bool	KeeperMainStandby(void);
 void	setupKeeperStandby(void);
 
-static void doPromote(void);
+static bool doPromote(void);
 static void doAfterCommand(void);
+static bool heartbeatServerStandby(int *retry_counts);
 
 /* GUC variables */
 char	*keeper_after_command;
 
 /* Variables for heartbeat */
-static int retry_count;
+static int *retry_counts;
 
 /*
  * Set up several parameters for standby mode.
@@ -45,24 +47,15 @@ static int retry_count;
 void
 setupKeeperStandby()
 {
-	PGconn *con;
-
-	/* Set up variables */
-	retry_count = 0;
-
-	/* Connection confirm */
-	if(!(con = PQconnectdb(KeeperMaster)))
-		ereport(ERROR,
-				(errmsg("could not establish connection to primary server : %s",
-						KeeperMaster)));
-
-	PQfinish(con);
+	/* Initialize */
+	retry_counts = resetRetryCounts(retry_counts);
 
 	/* Set process display which is exposed by ps command */
 	current_status = KEEPER_STANDBY_CONNECTED;
-	set_ps_display(getStatusPsString(current_status), false);
 
-	return;
+	/* Initialize own cache if pg_keeper is already installed */
+	if (checkExtensionInstalled())
+		updateLocalCache(false);
 }
 
 /*
@@ -101,27 +94,42 @@ KeeperMainStandby(void)
 			parse_synchronous_standby_names();
 		}
 
-		/*
-		 * Pooling to master server. If heartbeat is failed,
-		 * increment retry_count..
-		 */
-		if (!heartbeatServer(KeeperMaster, retry_count))
-			retry_count++;
-		else
-			retry_count = 0; /* reset count */
-
-		/*
-		 * If retry_count is reached to keeper_keepalives_count,
-		 * do promote the standby server to master server, and exit.
-		 */
-		if (retry_count >= keeper_keepalives_count)
+		/* If got SIGUSR1, update local cache for KeeperRepNodes */
+		if (got_sigusr1)
 		{
-			doPromote();
+			got_sigusr1 = false;
+			pg_usleep(5 * 1000L * 100L);
 
-			/* If after command is given, execute it */
-			if (keeper_after_command)
-				doAfterCommand();
+			/* Update own memeory */
+			updateLocalCache(false);
 
+			/* Update retry_counts information */
+			retry_counts = resetRetryCounts(retry_counts);
+		}
+
+		/*
+		 * Pooling to master server. If heartbeat is failed, increment retry_count.
+		 * As a result of polling, if retry_count is reached to
+		 * keeper_keepalives_count, do promote the standby server to master server,
+		 * and exit.
+		 */
+		if (!heartbeatServerStandby(retry_counts))
+		{
+			bool promoted;
+
+			/* Promote */
+			promoted = doPromote();
+
+			if (promoted)
+			{
+				/* If after command is given, execute it */
+				if (keeper_after_command)
+					doAfterCommand();
+			}
+
+			/* Change to status of this node to master mode */
+			current_status = KEEPER_MASTER_READY;
+			updateLocalCache(false);
 			return true;
 		}
 	}
@@ -134,29 +142,39 @@ KeeperMainStandby(void)
  * pg_ctl client tool. Put trigger file into $PGDATA, and send
  * SIGUSR1 signal to standby server.
  */
-static void
+static bool
 doPromote(void)
 {
 	char trigger_filepath[MAXPGPATH];
 	FILE *fp;
+	bool promote;
 
-	/* Create promote file newly */
-    snprintf(trigger_filepath, 1000, "%s/promote", DataDir);
-	if ((fp = fopen(trigger_filepath, "w")) == NULL)
-		ereport(ERROR,
-				(errmsg("could not create promote file: \"%s\"", trigger_filepath)));
+	promote = isNextMaster(keeper_node_name);
 
-	if (fclose(fp))
-		ereport(ERROR,
-				(errmsg("could not close promote file: \"%s\"", trigger_filepath)));
+	if (promote)
+	{
+		/* Create promote file newly */
+		snprintf(trigger_filepath, 1000, "%s/promote", DataDir);
+		if ((fp = fopen(trigger_filepath, "w")) == NULL)
+			ereport(ERROR,
+					(errmsg("could not create promote file: \"%s\"", trigger_filepath)));
 
-	elog(LOG,"promote standby server to primary server");
+		if (fclose(fp))
+			ereport(ERROR,
+					(errmsg("could not close promote file: \"%s\"", trigger_filepath)));
 
-	/* Do promote */
-	if (kill(PostmasterPid, SIGUSR1) != 0)
-		ereport(ERROR,
-				(errmsg("failed to send SIGUSR1 signal to postmaster process : %d",
-						PostmasterPid)));
+		elog(LOG,"promote standby server to primary server");
+
+		/* Do promote */
+		if (kill(PostmasterPid, SIGUSR1) != 0)
+			ereport(ERROR,
+					(errmsg("failed to send SIGUSR1 signal to postmaster process : %d",
+							PostmasterPid)));
+
+		return true;
+	}
+
+	return false;
 }
 
 /*
@@ -183,3 +201,98 @@ doAfterCommand(void)
 	}
 }
 
+/*
+ * heartbeatServerStandby()
+ * Polling to master server directly and indirectly via other standbys. Return false
+ * iif we could not poll to master via all standbys including itself.
+ */
+static bool
+heartbeatServerStandby(int *retry_counts)
+{
+#define KEEPER_SQL_INDIRECT_POOLING "SELECT pgkeeper.indirect_polling('%s')"
+	int i;
+	char *master_conninfo;
+	bool retry_count_reached = false;
+
+	/* Get master server connection information */
+	for (i = 0; i < nKeeperRepNodes; i++)
+	{
+		KeeperNode *node = &(KeeperRepNodes[i]);
+
+		if (node->is_master)
+		{
+			master_conninfo = node->conninfo;
+			break;
+		}
+	}
+
+	/*
+	 * Polling to the all servers. Return if pg_keeper made a dicision to
+	 * not be able to continue steaming replication. The standby server always
+	 * polling to master via all standby server indirectly. If all result of polling
+	 * via all standbys is false more then keeper_keepalives_count in a row the
+	 * we decide to promote. That's a our promoting policy.
+	 */
+	for (i = 0; i < nKeeperRepNodes; i++)
+	{
+		KeeperNode *node = &(KeeperRepNodes[i]);
+		char *connstr = node->conninfo;
+		bool ret;
+		char sql[BUFSIZE];
+		bool indirect_ret;
+
+		/*
+		 * We are not insterested in master server directly beacause
+		 * polling to master will be executed by execute indirectly
+		 * polling by itself.
+		 */
+		if (node->is_master)
+			continue;
+
+		/* Polling to master indirectly via other standby including itself */
+		snprintf(sql, BUFSIZE, KEEPER_SQL_INDIRECT_POOLING, master_conninfo);
+		ret = execSQL(connstr, sql, &indirect_ret);
+
+		if (!ret)
+		{
+			/* Emit warning log */
+			ereport(LOG,
+					(errmsg("pg_keeper failed to poll to \"%s\" at %d time(s)",
+							connstr, retry_counts[i])));
+
+			/* Neighbor standby migit be not available, ignore this result */
+			continue;
+		}
+
+		if (!indirect_ret)
+		{
+			/* Neighbor standby says that the master server might be not available */
+			(retry_counts[i])++;
+
+			/* Emit warning log */
+			ereport(LOG,
+					(errmsg("pg_keeper could not poll to the master server via \"%s\" at %d time(s)",
+							connstr, retry_counts[i])));
+
+			/* Check if retry_counts exceeds the threshold */
+			if (retry_counts[i] > keeper_keepalives_count)
+				retry_count_reached = true;
+			else
+				retry_count_reached = false;
+
+			continue;
+		}
+
+		/* Success to connect to the master indirectly */
+		retry_counts[i] = 0;
+	}
+
+	/*
+	 * retry_count_reached is true, which means this standby could not connect not only
+	 * the master but also other standbys could not connect to master server as well.
+	 */
+	if (retry_count_reached)
+		return false;
+
+	return true;
+}
